refactor(ui): address PR review comments on Icon component
- Move IconProps to Icon.interface.ts per repo convention (greptile
  comment: prop interfaces belong in dedicated .interface.ts files)
- Restore size=40 for image URLs in Certification.component.tsx —
  original 1.13 code used size=28 for named SVG icons and size=40 for
  image URLs; our migration collapsed both to 28 incorrectly

The gitar-bot className comment is already resolved: className lands on
<img> (not <span>) since the span-size fix commit.
