+++ openmetadata-ui/src/main/resources/ui/src/components/common/Icon/Icon.tsx
@@ -11,23 +11,9 @@
  *  limitations under the License.
  */
 import { Skeleton } from '@openmetadata/ui-core-components';
-import { FC, ReactNode, useEffect, useRef, useState } from 'react';
+import { FC, useEffect, useRef, useState } from 'react';
 import { getTagImageSrc, ICON_MAP, isImageUrl } from '../../../utils/IconUtils';
-
-export interface IconProps {
-  iconValue: string | undefined;
-  size?: number;
-  className?: string;
-  /** Layout/positioning styles (e.g. margin, flexShrink). Applied to the element
-   * occupying space in the caller's layout, regardless of loading state. */
-  wrapperStyle?: React.CSSProperties;
-  /** Cosmetic styles for the rendered icon/image itself (e.g. borderRadius). Never
-   * applied to the loading skeleton. */
-  imageStyle?: React.CSSProperties;
-  strokeWidth?: number;
-  alt?: string;
-  fallback?: ReactNode;
-}
+import { IconProps } from './Icon.interface';
 
 type IconLoadState = 'loading' | 'loaded' | 'error';
